refactor : RouteFinder를 유틸성 객체로
