+++ src/main/java/wooteco/subway/path/application/PathService.java
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Service;
 import wooteco.subway.line.application.LineService;
 import wooteco.subway.member.domain.MemberType;
 import wooteco.subway.path.domain.FarePolicy;
+import wooteco.subway.path.domain.RouteFinder;
 import wooteco.subway.path.domain.SubwayRoute;
 import wooteco.subway.path.dto.PathResponse;
 import wooteco.subway.station.application.StationService;
@@ -16,12 +17,10 @@ import java.util.List;
 public class PathService {
     private LineService lineService;
     private StationService stationService;
-    private RouteFinder routeFinder;
 
-    public PathService(LineService lineService, StationService stationService, RouteFinder routeFinder) {
+    public PathService(LineService lineService, StationService stationService) {
         this.lineService = lineService;
         this.stationService = stationService;
-        this.routeFinder = routeFinder;
     }
 
     public PathResponse findPath(Long sourceId, Long targetId, MemberType memberType) {
@@ -34,7 +33,7 @@ public class PathService {
         try {
             Station source = stationService.findStationById(sourceId);
             Station target = stationService.findStationById(targetId);
-            return routeFinder.find(lineService.findLines(), source, target);
+            return RouteFinder.find(lineService.findLines(), source, target);
         } catch (Exception e) {
             throw new PathException("적절하지 않은 구간의 경로 탐색입니다.");
         }
