+++ src/main/java/wooteco/subway/path/domain/RouteFinder.java
@@ -1,20 +1,16 @@
-package wooteco.subway.path.application;
+package wooteco.subway.path.domain;
 
 import org.jgrapht.GraphPath;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
-import org.springframework.stereotype.Service;
 import wooteco.subway.line.domain.Line;
-import wooteco.subway.path.domain.SectionEdge;
-import wooteco.subway.path.domain.SubwayGraph;
-import wooteco.subway.path.domain.SubwayRoute;
+import wooteco.subway.path.application.PathException;
 import wooteco.subway.station.domain.Station;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
-@Service
 public class RouteFinder {
-    public SubwayRoute find(List<Line> lines, Station source, Station target) {
+    public static SubwayRoute find(List<Line> lines, Station source, Station target) {
         if (source.equals(target)) {
             throw new PathException("적절하지 않은 구간의 경로 탐색입니다.");
         }
@@ -31,7 +27,7 @@ public class RouteFinder {
         return convertSubwayPath(path);
     }
 
-    private SubwayRoute convertSubwayPath(GraphPath graphPath) {
+    private static SubwayRoute convertSubwayPath(GraphPath graphPath) {
         List<SectionEdge> edges = (List<SectionEdge>) graphPath.getEdgeList()
                 .stream()
                 .collect(Collectors.toList());
