+++ src/main/resources/log4j2.xml
@@ -2,15 +2,18 @@
 <Configuration status="INFO">
     <Properties>
         <Property name="logNm">Application Log</Property>
-        <Property name="layoutPattern">MyCustomError %style{%d{yyyy/MM/dd HH:mm:ss,SSS}}{cyan} %highlight{[%-5p]}{FATAL=bg_red, ERROR=red,
-            INFO=green, DEBUG=blue} [%C] %style{[%t]}{yellow}- %m%n -</Property>
+        <Property name="layoutPattern">MyCustomError %style{%d{yyyy/MM/dd HH:mm:ss,SSS}}{cyan}
+            %highlight{[%-5p]}{FATAL=bg_red, ERROR=red,
+            INFO=green, DEBUG=blue} [%C] %style{[%t]}{yellow}- %m%n -
+        </Property>
     </Properties>
 
     <Appenders>
         <Console name="Console_Appender" target="SYSTEM_OUT">
             <PatternLayout pattern="${layoutPattern}"/>
         </Console>
-        <RollingFile name="File_Appender" fileName="logs/${logNm}.log" filePattern="logs/${logNm}_%d{yyyy-MM-dd}_%i.log.gz">
+        <RollingFile name="File_Appender" fileName="logs/${logNm}.log"
+                     filePattern="logs/${logNm}_%d{yyyy-MM-dd}_%i.log.gz">
             <PatternLayout pattern="${layoutPattern}"/>
             <Policies>
                 <SizeBasedTriggeringPolicy size="200KB"/>
@@ -27,15 +30,15 @@
         </Root>
         <Logger name="org.springframework" level="DEBUG"
                 additivity="false">
-            <AppenderRef ref="Console_Appender" />
+            <AppenderRef ref="Console_Appender"/>
             <AppenderRef ref="File_Appender"/>
         </Logger>
         <Logger name="com.fucct" level="INFO" additivity="false">
-            <AppenderRef ref="Console_Appender" />
+            <AppenderRef ref="Console_Appender"/>
             <AppenderRef ref="File_Appender"/>
         </Logger>
         <Logger name="com.fucct.springlog4j2.loggertest" level="TRACE" additivity="false">
-            <AppenderRef ref="Console_Appender" />
+            <AppenderRef ref="Console_Appender"/>
         </Logger>
     </Loggers>
 </Configuration>
